[Fix] #30- BaseResponse 수정
