+++ Winey/Winey/Source/Network/Base/BaseResponse.swift
@@ -7,13 +7,49 @@
 
 import Foundation
 
-struct BaseResponse<T: Codable>: Codable {
-    let status: Int
-    let success: Bool
-    let message: String
-    let data: T?
+struct BaseResponse<T: Decodable>: Decodable {
+    var status: Int
+    var success: Bool
+    var message: String?
+    var data: T?
+    
+    enum CodingKeys: String, CodingKey {
+        case status
+        case success
+        case message
+        case data
+    }
+    
+    init(from decoder: Decoder) throws {
+        let values = try decoder.container(keyedBy: CodingKeys.self)
+        status = (try? values.decode(Int.self, forKey: .status)) ?? 0
+        success = (try? values.decode(Bool.self, forKey: .success)) ?? false
+        message = (try? values.decode(String.self, forKey: .message)) ?? ""
+        data = (try? values.decode(T.self, forKey: .data)) ?? nil
+    }
 }
 
-/// data가 없는 API 통신에서 사용할 BlankData 구조체
-struct BlankData: Codable {
+struct BaseArrayResponse<T: Decodable>: Decodable {
+    let status: Int
+    let message: String?
+    let success: Bool?
+    let data: [T]?
+    
+    enum CodingKeys: String, CodingKey {
+        case message
+        case data
+        case status
+        case success
+    }
+    
+    init(from decoder: Decoder) throws {
+        let values = try decoder.container(keyedBy: CodingKeys.self)
+        status = (try? values.decode(Int.self, forKey: .status)) ?? 0
+        success = (try? values.decode(Bool.self, forKey: .success)) ?? false
+        message = (try? values.decode(String.self, forKey: .message)) ?? ""
+        data = (try? values.decode([T].self, forKey: .data)) ?? []
+    }
 }
+
+/// status, message, success 이외에 정보를 사용하지 않는 경우에 VoidType를 설정
+struct EmptyResponse: Decodable {}
