+++ Winey/Winey/Source/Network/Service/Feed/FeedService.swift
@@ -15,16 +15,16 @@ final class FeedService {
     
     init() { }
     
-    private(set) var totalFeedData: GeneralResponse<TotalFeedResponse>?
+    private(set) var totalFeedData: BaseResponse<TotalFeedResponse>?
     
     // 1, 전체 피드 조회하기
     
-    func getTotalFeed(page: Int, completion: @escaping (GeneralResponse<TotalFeedResponse>?) -> Void) {
+    func getTotalFeed(page: Int, completion: @escaping (BaseResponse<TotalFeedResponse>?) -> Void) {
         feedProvider.request(.getTotalFeed(page: page)) { [self] (result) in
             switch result {
             case .success(let response):
                 do {
-                    self.totalFeedData = try response.map(GeneralResponse<TotalFeedResponse>.self)
+                    self.totalFeedData = try response.map(BaseResponse<TotalFeedResponse>.self)
                     completion(totalFeedData)
                 } catch let error {
                     print(error.localizedDescription, 500)
